+++ functions/Get-TibberConsumption.ps1
@@ -8,6 +8,8 @@
         $response = Get-TibberConsumption -Last 10
         $maxCons = $response | Sort-Object -Property consumption -Descending | Select-Object -First 1
         Write-Host "Max power consumption $($maxCons.cost) $($maxCons.currency) ($($maxCons.consumption) $($maxCons.consumptionUnit) at $($maxCons.unitPrice)): $(([DateTime]$maxCons.from).ToString('HH:mm')) - $(([DateTime]$maxCons.to).ToString('HH:mm on yyyy-MM-dd'))"
+    .Example
+        Get-TibberConsumption -From ([DateTime]::Now).AddHours(-10)
     .Link
         Invoke-TibberQuery
     .Link
@@ -24,6 +26,14 @@
         [Parameter(ValueFromPipelineByPropertyName)]
         [ValidateSet('HOURLY', 'DAILY', 'WEEKLY', 'MONTHLY', 'ANNUAL')]
         [string] $Resolution = 'HOURLY',
+
+        # Specifies the start date for nodes to include in results.
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [DateTime] $From,
+
+        # Specifies the end date for nodes to include in results.
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [DateTime] $To = [DateTime]::Now,
 
         # Specifies the number of nodes to include in results, counting back from the latest entry.
         [Parameter(ValueFromPipelineByPropertyName)]
@@ -57,12 +67,31 @@
         if ($PSCmdlet.ParameterSetName -eq 'HomeId') {
             $homeNode = 'home'
             $query += "$homeNode(id:`"$HomeId`"){ "
-        }
-        else {
+        } else {
             $homeNode = 'homes'
             $query += "$homeNode{ "
         }
-        $arguments = "resolution:$Resolution, last:$Last"
+        if ($From) {
+            # date range provided, return nodes within the range
+            $From = $From.ToUniversalTime()
+            $splat = @{
+                Hour   = ($Resolution -ne 'HOURLY' ? 0 : $From.Hour)
+                Minute = 0
+                Second = 0
+            }
+            $after = Get-Date -Date $From @splat
+            $afterBase64 = [System.Convert]::ToBase64String( `
+                ([System.Text.Encoding]::UTF8.GetBytes( `
+                        (Get-Date -Date $after -Format s) `
+                    ) `
+                ) `
+            )
+            [int]$first = ($To.ToUniversalTime() - $From).TotalHours
+            $arguments = "resolution:$Resolution, after:`"$afterBase64`", first:$first"
+        } else {
+            # no date range provided, return the specified number of nodes
+            $arguments = "resolution:$Resolution, last:$Last"
+        }
         if ($FilterEmptyNodes.IsPresent) { $arguments += ", filterEmptyNodes:true" }
         $query += "consumption($arguments){ nodes{ $($Fields -join ','),__typename } }"
         $query += "}}}" # close query